requirements.txt
# Document Parsing
python-docx==1.2.0
pdfplumber==0.11.9

# AI Providers
anthropic>=0.42.0
openai>=1.50.0
google-genai>=1.0.0
mistralai>=1.0.0
cohere>=5.0.0

# GUI
customtkinter==5.2.2

# Configuration
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Packaging
pyinstaller==6.11.1
